:hammer:Refact: view와 로직 분리
view와 로직 분리
