+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { Route, Routes } from "react-router-dom";
 import RootPage from "./routes/RootPage";
-import LoginPage from "./routes/LoginPage";
 import SignUpPage from "./routes/SignUpPage";
 import SignUpOptionsPage from "./routes/SignUpOptionPage";
 import KakaoPage from "./routes/KakaoPage";
@@ -15,7 +14,6 @@ function App() {
     <>
       <Routes>
         <Route path="" element={<RootPage />} />
-        <Route path="login" element={<LoginPage />} />
         <Route path="kakao" element={<KakaoPage />} />
         <Route path="signup">
           <Route path="" element={<SignUpOptionsPage />} />
@@ -24,7 +22,7 @@ function App() {
           <Route path="success" element={<SignUpSuccessPage />} />
         </Route>
         <Route path="analyze" element={<AnalyzePage />} />
-        <Route path="mathcing" element={<MatchingPage />} />
+        <Route path="matching" element={<MatchingPage />} />
         <Route path="mypage" element={<MyPage />} />
         <Route path="connect" element={<ConnectPage />} />
       </Routes>
