+++ src/Components/SignUp/SignUpForm.tsx
@@ -1,155 +1,21 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import request from "@/libs/axios";
-import { kakaoPostOnCompleteData } from "@/types/kakaoPostOnCompleteData";
+import useSignUpForm from "@/Logic/Components/SignUp/useSignUpForm";
 
-interface infoObj {
-  email: string;
-  password: string;
-  passwordVerification: string;
-  name: string;
-  birthDate: Date;
-  gender: string;
-  phone1: string;
-  phone2: string;
-  phone3: string;
-  address: string;
-  postcode: string;
-  detailAddress: string;
-  extraAddress: string;
-}
-
-const { daum } = window;
-
-const SignUpForm: React.FC = () => {
-  const [info, setInfo] = useState<infoObj>({
-    email: "",
-    password: "",
-    passwordVerification: "",
-    name: "",
-    birthDate: new Date(),
-    gender: "",
-    phone1: "",
-    phone2: "",
-    phone3: "",
-    address: "",
-    postcode: "",
-    detailAddress: "",
-    extraAddress: "",
-  });
-
-  const detailAddressRef = useRef(null);
-  const navigate = useNavigate();
-  const [duplicationCheckMent, setDuplicateCheckMent] = useState<string>("");
-
-  const signUpHandler = useCallback(
-    async (event: React.FormEvent<HTMLFormElement>) => {
-      event.preventDefault();
-      console.log(info);
-      try {
-        const res = await request.post(`/auth/register`, {
-          name: info.name,
-          email: info.email,
-          gender: info.gender,
-          birth: info.birthDate,
-          phone: `${info.phone1}-${info.phone2}-${info.phone3}`,
-          address: `${info.address} ${info.detailAddress} ${info.extraAddress}`,
-          pw: info.password,
-        });
-        console.log(res);
-        navigate("/signup/success");
-      } catch (error) {
-        console.log("[에러]", error);
-      }
-    },
-    [navigate, info]
-  );
-
-  const changeHandler = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => {
-      setInfo((state) => {
-        return {
-          ...state,
-          [event.target.name]: event.target.value,
-        };
-      });
-    },
-    []
-  );
-
-  useEffect(() => {
-    console.log(info);
-  }, [info]);
-
-  const addressHandler = useCallback(() => {
-    new daum.Postcode({
-      oncomplete: function (data: kakaoPostOnCompleteData) {
-        // 팝업에서 검색결과 항목을 클릭했을때 실행할 코드를 작성하는 부분.
-
-        // 각 주소의 노출 규칙에 따라 주소를 조합한다.
-        // 내려오는 변수가 값이 없는 경우엔 공백('')값을 가지므로, 이를 참고하여 분기 한다.
-        let addr = ""; // 주소 변수
-        let extraAddr = ""; // 참고항목 변수
-
-        //사용자가 선택한 주소 타입에 따라 해당 주소 값을 가져온다.
-        if (data.userSelectedType === "R") {
-          // 사용자가 도로명 주소를 선택했을 경우
-          addr = data.roadAddress;
-        } else {
-          // 사용자가 지번 주소를 선택했을 경우(J)
-          addr = data.jibunAddress;
-        }
-
-        // 사용자가 선택한 주소가 도로명 타입일때 참고항목을 조합한다.
-        if (data.userSelectedType === "R") {
-          // 법정동명이 있을 경우 추가한다. (법정리는 제외)
-          // 법정동의 경우 마지막 문자가 "동/로/가"로 끝난다.
-          if (data.bname !== "" && /[동|로|가]$/g.test(data.bname)) {
-            extraAddr += data.bname;
-          }
-          // 건물명이 있고, 공동주택일 경우 추가한다.
-          if (data.buildingName !== "" && data.apartment === "Y") {
-            extraAddr +=
-              extraAddr !== "" ? ", " + data.buildingName : data.buildingName;
-          }
-          // 표시할 참고항목이 있을 경우, 괄호까지 추가한 최종 문자열을 만든다.
-          if (extraAddr !== "") {
-            extraAddr = " (" + extraAddr + ")";
-          }
-          // 조합된 참고항목을 해당 필드에 넣는다.
-        }
-
-        setInfo((state) => ({
-          ...state,
-          postcode: data.zonecode,
-          address: addr,
-          extraAddress: extraAddr,
-        }));
-
-        if (detailAddressRef.current !== null) {
-          (detailAddressRef.current as HTMLInputElement).focus();
-        }
-      },
-    }).open();
-  }, []);
-
-  const checkEmailDuplication = useCallback(async () => {
-    try {
-      const res = await request.get(`/auth/dupcheck?email=${info.email}`);
-      if (res.data.result) {
-        setDuplicateCheckMent("사용 가능합니다.");
-      } else {
-        setDuplicateCheckMent("이미 가입된 계정입니다.");
-      }
-    } catch (error) {
-      console.log("[에러]", error);
-    }
-  }, [info.email]);
+const SignUpForm = () => {
+  const {
+    info,
+    signUpUser,
+    handleInputChange,
+    checkEmailDuplication,
+    handleAddress,
+    checkButtonClicked,
+    isDuplicated,
+    detailAddressRef,
+  } = useSignUpForm();
 
   return (
     <form
       className="[&_input]:border-b-[1px] [&_input]:border-black [&_input]:outline-none"
-      onSubmit={signUpHandler}
+      onSubmit={signUpUser}
     >
       <div>
         <label htmlFor="email">이메일</label>
@@ -158,20 +24,24 @@ const SignUpForm: React.FC = () => {
           id="email"
           name="email"
           value={info.email}
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
         <button type="button" onClick={checkEmailDuplication}>
           중복 확인
         </button>
       </div>
-      <div>{duplicationCheckMent}</div>
+      {checkButtonClicked && isDuplicated ? (
+        <div>이미 사용중인 이메일입니다.</div>
+      ) : (
+        <div>사용 가능합니다.</div>
+      )}
       <div>
         <label htmlFor="password">비밀번호</label>
         <input
           type="password"
           id="password"
           name="password"
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
       </div>
       <div>
@@ -180,16 +50,16 @@ const SignUpForm: React.FC = () => {
           type="password"
           id="passwordVerification"
           name="passwordVerification"
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
       </div>
       <div>
         <label htmlFor="name">이름</label>
-        <input type="text" name="name" id="name" onChange={changeHandler} />
+        <input type="text" name="name" id="name" onChange={handleInputChange} />
       </div>
       <div>
         <label htmlFor="name">생년월일</label>
-        <input type="date" name="birthDate" onChange={changeHandler} />
+        <input type="date" name="birthDate" onChange={handleInputChange} />
       </div>
       <div>
         <div>성별</div>
@@ -199,7 +69,7 @@ const SignUpForm: React.FC = () => {
           name="gender"
           value="male"
           required
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
         <label htmlFor="male">남자</label>
         <input
@@ -207,7 +77,7 @@ const SignUpForm: React.FC = () => {
           id="female"
           name="gender"
           value="female"
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
         <label htmlFor="female">여자</label>
       </div>
@@ -217,7 +87,7 @@ const SignUpForm: React.FC = () => {
           type="number"
           id="phone1"
           name="phone1"
-          onChange={changeHandler}
+          onChange={handleInputChange}
           required
         />
         <span>-</span>
@@ -226,7 +96,7 @@ const SignUpForm: React.FC = () => {
           id="phone2"
           name="phone2"
           value={info.phone2}
-          onChange={changeHandler}
+          onChange={handleInputChange}
           required
         />
         <span>-</span>
@@ -235,7 +105,7 @@ const SignUpForm: React.FC = () => {
           id="phone3"
           name="phone3"
           value={info.phone3}
-          onChange={changeHandler}
+          onChange={handleInputChange}
           required
         />
       </div>
@@ -245,9 +115,9 @@ const SignUpForm: React.FC = () => {
           name="postcode"
           id="postcode"
           value={info.postcode}
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
-        <button type="button" onClick={addressHandler}>
+        <button type="button" onClick={handleAddress}>
           찾기
         </button>
       </div>
@@ -256,7 +126,7 @@ const SignUpForm: React.FC = () => {
           placeholder="도로명 주소"
           name="address"
           value={info.address}
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
       </div>
       <div>
@@ -264,7 +134,7 @@ const SignUpForm: React.FC = () => {
           placeholder="상세 주소"
           name="detailAddress"
           ref={detailAddressRef}
-          onChange={changeHandler}
+          onChange={handleInputChange}
         />
       </div>
       <div>
