+++ src/routes/AnalyzePage.tsx
@@ -1,4 +1,4 @@
-const AnalyzePage: React.FC = () => {
+const AnalyzePage = () => {
   return <h1>분석 페이지</h1>;
 };
 export default AnalyzePage;
