+++ src/routes/ConnectPage.tsx
@@ -1,4 +1,4 @@
-const ConnectPage: React.FC = () => {
+const ConnectPage = () => {
   return (
     <div className="fixed top-0 left-0 w-[100vw] h-[100vh] bg-white [&_input]:border-b [&_input]:border-black [&_input]:outline-none">
       <h1>네트워크 연결</h1>
