--- src/routes/LoginPage.tsx
@@ -1,5 +1,0 @@
-const LoginPage: React.FC = () => {
-  return <h1>로그인</h1>;
-};
-
-export default LoginPage;
