+++ src/routes/MyPage.tsx
@@ -1,4 +1,4 @@
-const MyPage: React.FC = () => {
+const MyPage = () => {
   return <h1>마이 페이지</h1>;
 };
