+++ src/routes/RootPage.tsx
@@ -1,4 +1,4 @@
-const RootPage: React.FC = () => {
+const RootPage = () => {
   return (
     <>
       <h1 className="text-3xl font-bold underline">파트라슈</h1>
