+++ src/routes/SignUpPage.tsx
@@ -1,7 +1,7 @@
 import { useLocation } from "react-router-dom";
 import SignUpForm from "../Components/SignUp/SignUpForm";
 
-const SignUpPage: React.FC = () => {
+const SignUpPage = () => {
   const location = useLocation();
   console.log("location.state", location.state);
